pyproject.toml
[project]
name = "capstone-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]>=0.135.3",
  "pydantic-settings>=2.13.1",
  "sqlalchemy>=2.0.49",
  "uvicorn>=0.44.0",
  "langchain>=0.1.0",
  "langgraph>=0.0.21",
  "langchain-google-genai>=0.0.9",
  "psycopg2-binary>=2.9.11",
  "pyjwt>=2.12.1",
  "bcrypt==4.3.0",
]
[tool.pyright]
venvPath = "."
venv = ".venv"
extraPaths = ["./app"]
